requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
certifi==2025.8.3
click==8.2.1
cloudinary==1.44.1
Deprecated==1.2.18
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
limits==5.5.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
wrapt==1.17.3
